refactor dpm campaign readiness routes
